refactor + more tests
