+++ cypress/e2e/settings/menu.cy.ts
@@ -37,10 +37,27 @@ describe('Menu configuration', () => {
     cy.get('#link-title').type('Link 3', { delay: 0 });
     cy.get('#link-url').type('www.exmple.com', { delay: 0 });
     cy.getByTestId('menu-form-submit').click();
+  });
 
+  it('should alert users of unsaved changes', () => {
+    cy.contains('a', 'Account').click();
+    cy.get('[data-testid="modal"]').within(() => {
+      cy.contains('You have unsaved changes. Do you want to continue?');
+      cy.contains('button', 'Cancel').click();
+    });
+  });
+
+  it('should save', () => {
     cy.getByTestId('menu-save').click();
     cy.contains('Dismiss').click();
     cy.wait('@fetchLinks');
+    cy.getByTestId('menu-save').should('be.disabled');
+  });
+
+  it('should not should the unsaved changes alert', () => {
+    cy.contains('a', 'Account').click();
+    cy.contains('a', 'Menu').click();
+    cy.contains('caption', 'Menu');
   });
 
   it('tests Add groups', () => {
@@ -66,9 +83,21 @@ describe('Menu configuration', () => {
     cy.get('#link-title').type(' edited', { delay: 0 });
     cy.get('#link-group').select('Group 2');
     cy.getByTestId('menu-form-submit').click();
+  });
+
+  it('should alert users of unsaved changes after editing', () => {
+    cy.contains('a', 'Account').click();
+    cy.get('[data-testid="modal"]').within(() => {
+      cy.contains('You have unsaved changes. Do you want to continue?');
+      cy.contains('button', 'Cancel').click();
+    });
+  });
+
+  it('should save the editied links', () => {
     cy.getByTestId('menu-save').click();
     cy.contains('Dismiss').click();
     cy.wait('@fetchLinks');
+    cy.getByTestId('menu-save').should('be.disabled');
   });
 
   it('tests edit groups', () => {
